refactor: render ingest source CSV through view
